Various Custom Event additions and API tweaks:
Added Events to call when a player issues a command or performs an
event, that can be cancelled by any listener. Changed the
BuildingMaterial enum to be a class that represents a Building Material,
a list of all different types of blocks is generated at load up by the
plugin, Materials can be accessed in a similar way as before. Added a
SQL API for accessing all SQL Server tables in the one Database, this
reduces the nunmber of connections to the SQL server down to one, and
allows for easy optimisation of the server accessor. Finally relocated
packages so that all classes are now under the main
me.Tiernanator.Utilities parent package.